use fetched txs
